soc: qcom: Add better support for early random numbers
The existing support for generating random numbers relied on
a hacked up version of CONFIG_ARCH_RANDOM and was prone to
scheduling while atomic bugs due to needing to take a mutex.
We don't actually need the additional randomness all the time,
the pool just needs to be initialized with some amount of random
data to avoid getting the same result each time. Switch to
this method for initializing the random pool.

Signed-off-by: Laura Abbott <lauraa@codeaurora.org>
Signed-off-by: Pranav Vashi <neobuddy89@gmail.com>