msm: vidc: Cancel pending power collapse works in core_release
There is a possiblity of power collapse kicking in between core_release
and unload_fw.  In the case that it does, we'd have to pull the hardware
out of PC in unload_fw.  However this isn't possible pulling out of PC
would re-allocate OCMEM and lead to an OCMEM leak prior to unloading the
firmware.

To avoid the unneccesary trouble, cancel the PC work in core_release so
that we guarentee the hardware to be in a good state in unload_fw.

Change-Id: Iae785322a230c8ce8bfcc0dfd2f4ba5ff16bf698
Signed-off-by: Deva Ramasubramanian <dramasub@codeaurora.org>
Signed-off-by: franciscofranco <franciscofranco.1990@gmail.com>